Bring back the Client, but autogenerated this time!
This commit enables us to actually match responses with requests. I realized that the ClientManager is actually useless as it's more comfortable to have a Client object, but autogenerated for all the needed methods. The last remaining part is serializing the response to the right type and make the methods actually return it (instead of returning an useless bool like we're doing now).